Updaing files and types
