+++ libs/ui/src/components/atoms/Card/index.ts
@@ -1,1 +1,5 @@
-export * from './Card';
+import { Card } from './Card';
+import type { CardProps } from './Card.types';
+
+export type { CardProps };
+export { Card };
