+++ libs/ui/src/components/atoms/Checkbox/Checkbox.tsx
@@ -1,11 +1,10 @@
-import React, { RefObject } from 'react';
-
+import React from 'react';
 import classNames from 'classnames';
 import { RiCheckLine, RiAsterisk } from 'react-icons/ri';
-import { CheckboxProps } from '@radix-ui/react-checkbox';
 
 import { Label } from '../Label/Label';
 import { Icon } from '../Icon';
+import type { CheckboxProps } from './Checkbox.types';
 import {
   StyledCheckbox,
   StyledIndicator,
@@ -14,36 +13,32 @@ import {
   RequiredAsterisk,
 } from './Checkbox.styles';
 
-type Ref =
-  | RefObject<HTMLButtonElement>
-  | ((instance: HTMLButtonElement | null) => void)
-  | null
-  | undefined;
+export const Checkbox = React.forwardRef<HTMLButtonElement, CheckboxProps>(
+  (props, ref) => {
+    const { id, title, disabled, required } = props;
 
-export const Checkbox = React.forwardRef((props: CheckboxProps, ref: Ref) => {
-  const { id, title, disabled, required } = props;
+    const classes = classNames({
+      disabled,
+    });
 
-  const classes = classNames({
-    disabled,
-  });
-
-  return (
-    <Container>
-      <StyledCheckbox {...props} ref={ref}>
-        <StyledIndicator className={classes}>
-          <RiCheckLine />
-        </StyledIndicator>
-      </StyledCheckbox>
-      <LabelContainer>
-        {required && (
-          <RequiredAsterisk>
-            <Icon label="Required">
-              <RiAsterisk />
-            </Icon>
-          </RequiredAsterisk>
-        )}
-        <Label id={id}>{title ? title : 'No Title Found'}</Label>
-      </LabelContainer>
-    </Container>
-  );
-});
+    return (
+      <Container>
+        <StyledCheckbox {...props} ref={ref}>
+          <StyledIndicator className={classes}>
+            <RiCheckLine />
+          </StyledIndicator>
+        </StyledCheckbox>
+        <LabelContainer>
+          {required && (
+            <RequiredAsterisk>
+              <Icon label="Required">
+                <RiAsterisk />
+              </Icon>
+            </RequiredAsterisk>
+          )}
+          <Label id={id}>{title ? title : 'No Title Found'}</Label>
+        </LabelContainer>
+      </Container>
+    );
+  }
+);
