+++ libs/ui/src/components/atoms/Checkbox/Checkbox.types.ts
@@ -1,0 +1,3 @@
+import { CheckboxProps } from '@radix-ui/react-checkbox';
+
+export type { CheckboxProps };
