+++ libs/ui/src/components/atoms/Checkbox/index.ts
@@ -1,1 +1,5 @@
-export * from './Checkbox';
+import { Checkbox } from './Checkbox';
+import type { CheckboxProps } from './Checkbox.types';
+
+export type { CheckboxProps };
+export { Checkbox };
