+++ libs/ui/src/components/atoms/HelperTexts/HelperText.types.ts
@@ -1,0 +1,11 @@
+import { IconType } from 'react-icons';
+
+export type HelperTextType = {
+  color?: string;
+  icon?: IconType;
+  children: React.ReactNode;
+};
+
+export type SpecialHelperText = {
+  children: React.ReactNode;
+};
