+++ libs/ui/src/components/atoms/Input/Input.types.ts
@@ -1,0 +1,9 @@
+import { IconType } from 'react-icons';
+
+import { Field } from '../../../types/formAndField';
+
+export type InputProps = Field & {
+  icon?: IconType;
+  onChange?: React.ChangeEventHandler<HTMLInputElement> | undefined;
+  className?: string;
+};
