+++ libs/ui/src/components/atoms/Label/Label.tsx
@@ -1,9 +1,5 @@
+import type { LabelProps } from './Label.types';
 import { StyledLabel } from './Label.styles';
-
-type LabelProps = {
-  id?: string;
-  children: React.ReactChild;
-};
 
 export const Label: React.FC<LabelProps> = ({
   children = 'label',
