+++ libs/ui/src/components/atoms/Label/Label.types.ts
@@ -1,0 +1,4 @@
+export type LabelProps = {
+  id?: string;
+  children: React.ReactChild;
+};
