+++ libs/ui/src/components/atoms/Label/index.ts
@@ -1,1 +1,5 @@
-export * from './Label';
+import { Label } from './Label';
+import type { LabelProps } from './Label.types';
+
+export type { LabelProps };
+export { Label };
